faas/text-keyword-extractor/requirements.txt
multi-rake
guess_language-spirit